rasa_python/rasa_v3/requirements.txt
absl-py==0.10.0
aiofiles==0.6.0
aiohttp==3.6.3
APScheduler==3.6.3
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
boto3==1.17.69
botocore==1.20.69
cachetools==4.1.1
certifi==2021.5.30
cffi==1.14.5
click==8.0.1
cloudpickle==1.4.1
colorclass==2.2.0
coloredlogs==14.3
colorhash==1.0.3
configparser==5.0.0
contextvars==2.4
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
et-xmlfile==1.0.1
fbmessenger==6.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
gensim==3.8.1
google==3.0.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.37.1
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.11.1
httplib2==0.18.1
httptools==0.2.0
httpx==0.15.4
huggingface-hub==0.0.8
humanfriendly==9.1
hyperframe==5.2.0
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.6.1
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.4.2
jsonschema==3.2.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libretranslatepy==2.1.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
multidict==4.7.6
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opencc-python-reimplemented==0.1.6
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==20.9
pika==1.1.0
Pillow==8.2.0
prompt-toolkit==2.0.10
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.2
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.1
python-telegram-bot==12.8
pytz==2020.4
PyYAML==5.4.1
questionary==1.5.2
rasa==2.0.2
rasa-sdk==2.6.0
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-API==1.9.1
rsa==4.6
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
sacremoses==0.0.45
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.23.2
scipy==1.4.1
sentry-sdk==0.19.5
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smart-open==3.0.0
sniffio==1.2.0
soupsieve==2.0.1
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.37.7
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.2
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.8.1
tornado==6.1
tqdm==4.41.0
transformers==4.6.1
translate==3.6.1
twilio==6.45.4
typeguard==2.12.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.26.6
waitress==1.4.4
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
yarl==1.5.1
zipp==3.5.0
